requirements.txt
bcrypt
celery
fastapi
pandas
passlib
pydantic
pydantic_core
python-jose
python-multipart
redis
SQLAlchemy
uvicorn0
